moved lines method code to linecollection
